+++ src/routes/links/links.json
@@ -2,7 +2,7 @@
 	"links": [
 		{
 			"title": "Blog",
-			"href": "https://jtabb.dev"
+			"href": "/"
 		},
 		{
 			"title": "GitHub",
