+++ static/robots.txt
@@ -3,5 +3,3 @@
 User-agent: *
 Disallow:
 
-User-agent: GPTBot
-Disallow: /
